pyproject.toml
[tool.poetry]
package-mode = false
name = "tesi-magistrale"
description = "Tesi magistrale di Mariano Mollo"
authors = ["Mariano Mollo <marian.mollo@studenti.unina.it>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.8"
ase = "^3.23.0"
mace-torch = {git = "https://github.com/ACEsuit/mace.git"}
torch-dftd = "^0.4.0"
phonopy = "^2.21.2"
seekpath = "^2.1.0"
marimo = "^0.7.0"
pandas = "^2.2.0"
polars = "^0.20.13"
jupyterlab = "^4.1.6"
tornado = "^6.4"
mdanalysis = "^2.7.0"
symfc = "^1.0.2"
scipy = "^1.13.1"
pint = "^0.24.1"
natsort = "^8.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
